+++ sdk/resourcemanager/network/armnetwork/constants.go
@@ -11,7 +11,7 @@ package armnetwork
 
 const (
 	moduleName    = "armnetwork"
-	moduleVersion = "v2.1.0"
+	moduleVersion = "v2.2.0"
 )
 
 // Access - Access to be allowed or denied.
@@ -79,6 +79,26 @@ func PossibleAdminRuleKindValues() []AdminRuleKind {
 	return []AdminRuleKind{
 		AdminRuleKindCustom,
 		AdminRuleKindDefault,
+	}
+}
+
+// AnalysisModes - Analysis Modes.
+type AnalysisModes string
+
+const (
+	AnalysisModesDefault  AnalysisModes = "Default"
+	AnalysisModesNCD      AnalysisModes = "NCD"
+	AnalysisModesNextHop  AnalysisModes = "NextHop"
+	AnalysisModesPortScan AnalysisModes = "PortScan"
+)
+
+// PossibleAnalysisModesValues returns the possible values for the AnalysisModes const type.
+func PossibleAnalysisModesValues() []AnalysisModes {
+	return []AnalysisModes{
+		AnalysisModesDefault,
+		AnalysisModesNCD,
+		AnalysisModesNextHop,
+		AnalysisModesPortScan,
 	}
 }
 
@@ -286,6 +306,7 @@ func PossibleApplicationGatewayRuleSetStatusOptionsValues() []ApplicationGateway
 type ApplicationGatewaySKUName string
 
 const (
+	ApplicationGatewaySKUNameStandardBasic  ApplicationGatewaySKUName = "Standard_Basic"
 	ApplicationGatewaySKUNameStandardLarge  ApplicationGatewaySKUName = "Standard_Large"
 	ApplicationGatewaySKUNameStandardMedium ApplicationGatewaySKUName = "Standard_Medium"
 	ApplicationGatewaySKUNameStandardSmall  ApplicationGatewaySKUName = "Standard_Small"
@@ -298,6 +319,7 @@ const (
 // PossibleApplicationGatewaySKUNameValues returns the possible values for the ApplicationGatewaySKUName const type.
 func PossibleApplicationGatewaySKUNameValues() []ApplicationGatewaySKUName {
 	return []ApplicationGatewaySKUName{
+		ApplicationGatewaySKUNameStandardBasic,
 		ApplicationGatewaySKUNameStandardLarge,
 		ApplicationGatewaySKUNameStandardMedium,
 		ApplicationGatewaySKUNameStandardSmall,
@@ -440,16 +462,18 @@ func PossibleApplicationGatewaySSLProtocolValues() []ApplicationGatewaySSLProtoc
 type ApplicationGatewayTier string
 
 const (
-	ApplicationGatewayTierStandard   ApplicationGatewayTier = "Standard"
-	ApplicationGatewayTierStandardV2 ApplicationGatewayTier = "Standard_v2"
-	ApplicationGatewayTierWAF        ApplicationGatewayTier = "WAF"
-	ApplicationGatewayTierWAFV2      ApplicationGatewayTier = "WAF_v2"
+	ApplicationGatewayTierStandard      ApplicationGatewayTier = "Standard"
+	ApplicationGatewayTierStandardBasic ApplicationGatewayTier = "Standard_Basic"
+	ApplicationGatewayTierStandardV2    ApplicationGatewayTier = "Standard_v2"
+	ApplicationGatewayTierWAF           ApplicationGatewayTier = "WAF"
+	ApplicationGatewayTierWAFV2         ApplicationGatewayTier = "WAF_v2"
 )
 
 // PossibleApplicationGatewayTierValues returns the possible values for the ApplicationGatewayTier const type.
 func PossibleApplicationGatewayTierValues() []ApplicationGatewayTier {
 	return []ApplicationGatewayTier{
 		ApplicationGatewayTierStandard,
+		ApplicationGatewayTierStandardBasic,
 		ApplicationGatewayTierStandardV2,
 		ApplicationGatewayTierWAF,
 		ApplicationGatewayTierWAFV2,